Refactor notification imports and adjust dependencies
