+++ pkg/platforms/mister/platform.go
@@ -4,7 +4,7 @@ package mister
 
 import (
 	"fmt"
-	"github.com/wizzomafizzo/tapto/pkg/service/notifications"
+	"github.com/wizzomafizzo/tapto/pkg/service/api/notifications"
 	"os"
 	"os/exec"
 	"regexp"
