+++ pkg/platforms/mistex/platform.go
@@ -4,7 +4,7 @@ package mistex
 
 import (
 	"fmt"
-	"github.com/wizzomafizzo/tapto/pkg/service/notifications"
+	"github.com/wizzomafizzo/tapto/pkg/service/api/notifications"
 	"os"
 	"os/exec"
 	"path/filepath"
