+++ pkg/platforms/windows/platform.go
@@ -2,7 +2,7 @@ package windows
 
 import (
 	"errors"
-	"github.com/wizzomafizzo/tapto/pkg/service/notifications"
+	"github.com/wizzomafizzo/tapto/pkg/service/api/notifications"
 	"os"
 	"os/exec"
 	"path/filepath"
