+++ pkg/service/state/state.go
@@ -1,7 +1,7 @@
 package state
 
 import (
-	"github.com/wizzomafizzo/tapto/pkg/service/notifications"
+	"github.com/wizzomafizzo/tapto/pkg/service/api/notifications"
 	"sync"
 
 	"github.com/rs/zerolog/log"
